requirement.txt
telethon >= 1.24.0
cryptg >= 0.3.1
gcloud-aio-storage >= 7.0.1